task_7. Done
