+++ app/src/main/java/com/example/aydar/editingusersprofile/MusicService.kt
@@ -8,13 +8,10 @@ import android.os.Binder
 import android.os.Build
 import android.os.IBinder
 import android.support.v4.app.NotificationCompat
-import android.widget.ImageView
-import kotlinx.android.synthetic.main.activity_music.*
 
 class MusicService : Service() {
     val CONST_ACTION_MAIN = "ACTION_MAIN"
     val CONST_ACTION_PLAY = "ACTION_PLAY"
-    val CONST_ACTION_PAUSE = "ACTION_PAUSE"
     val CONST_ACTION_NEXT = "ACTION_NEXT"
     val CONST_ACTION_PREV = "ACTION_PREV"
     var currentSong = 0
@@ -24,19 +21,31 @@ class MusicService : Service() {
     private var currentPosition: Int = 0
     private val playList: ArrayList<Song> = Song.songs
     private lateinit var listener: Callback
-    var imagePlay : ImageView = MainActivity().image_play
-
 
     override fun onBind(intent: Intent): IBinder {
         return mBinder
     }
 
-    inner class MyBinder : Binder() {
-        internal
-        val service: MusicService
-            get() = this@MusicService
+    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+        val songNum: Int? = intent?.getIntExtra(MainActivity.CONST_SONG_NUM, -1)
+        if (songNum != null && songNum >= 0) {
+            launchForeground()
+            playSpecificMusic(songNum)
+        }
+        if (intent?.action == CONST_ACTION_PLAY) {
+            if (mediaPlayer?.isPlaying!!)
+                pauseMusic()
+            else
+                playOrStopMusic()
+        }
+        if (intent?.action == CONST_ACTION_PREV) {
+            playPreviousSong()
+        }
+        if (intent?.action == CONST_ACTION_NEXT) {
+            playNextSong()
+        }
+        return START_NOT_STICKY;
     }
-
 
     fun playOrStopMusic() {
         if (mediaPlayer == null) {
@@ -46,15 +55,11 @@ class MusicService : Service() {
         } else if (mediaPlayer != null && !mediaPlayer!!.isPlaying) {
             mediaPlayer?.seekTo(currentPosition)
             mediaPlayer?.start()
-            //TODO поменять иконку stop на play
-            imagePlay.setImageResource(R.drawable.play)
         } else if (mediaPlayer != null && mediaPlayer!!.isPlaying) {
             mediaPlayer?.let {
                 it.pause()
                 currentPosition = it.currentPosition
             }
-            //TODO поменять иконку play на stop
-            imagePlay.setImageResource(R.drawable.pause);
         }
     }
 
@@ -63,7 +68,6 @@ class MusicService : Service() {
             it.pause()
             currentPosition = it.currentPosition
         }
-
     }
 
     fun playNextSong() {
@@ -85,20 +89,13 @@ class MusicService : Service() {
         }
     }
 
+    inner class MyBinder : Binder() {
+        internal
+        val service: MusicService
+            get() = this@MusicService
+    }
 
     private fun launchForeground() {
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val name = "Channel name"
-            val descriptionText = "Channel Description"
-            val importance = NotificationManager.IMPORTANCE_DEFAULT
-            val channel = NotificationChannel("channelid", name, importance).apply {
-                description = descriptionText
-            }
-            val notificationManager: NotificationManager =
-                    this.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(channel)
-        }
 
         val notificationIntent: Intent = Intent(this, MainActivity::class.java)
         notificationIntent.setAction(CONST_ACTION_MAIN)
@@ -117,7 +114,6 @@ class MusicService : Service() {
         nextIntent.setAction(CONST_ACTION_NEXT)
         val pNextIntent: PendingIntent = PendingIntent.getService(this, 0, nextIntent, 0)
 
-
         val notification: Notification = NotificationCompat.Builder(this)
                 .setContentTitle("Music PLayer")
                 .setTicker("MPlayer")
@@ -133,28 +129,6 @@ class MusicService : Service() {
         startForeground(123, notification)
     }
 
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        val songNum: Int? = intent?.getIntExtra(MainActivity.CONST_SONG_NUM, -1)
-        if (songNum != null && songNum >= 0) {
-
-            launchForeground()
-            playSpecificMusic(songNum)
-        }
-        if (intent?.action == CONST_ACTION_PLAY) {
-            if (mediaPlayer?.isPlaying!!)
-                pauseMusic()
-            else
-                playOrStopMusic()
-        }
-        if (intent?.action == CONST_ACTION_PREV) {
-            playPreviousSong()
-        }
-        if (intent?.action == CONST_ACTION_NEXT) {
-            playNextSong()
-        }
-        return START_NOT_STICKY;
-    }
-
     private fun playSpecificMusic(position: Int) {
         mediaPlayer?.stop()
         mediaPlayer = MediaPlayer.create(this, playList[position].songResourse)
@@ -168,5 +142,9 @@ class MusicService : Service() {
 
     interface Callback {
         fun changeData(p: Int)
+    }
+
+    fun isPlayerPlaying(): Boolean {
+        return mediaPlayer != null && mediaPlayer?.isPlaying!!
     }
 }
