+++ app/src/main/java/com/example/aydar/editingusersprofile/SettingsActivity.kt
@@ -5,25 +5,19 @@ import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.preference.Preference
 import android.preference.PreferenceActivity
-import android.support.v7.preference.PreferenceManager
-import android.util.Log
 
 class SettingsActivity : PreferenceActivity(), SharedPreferences.OnSharedPreferenceChangeListener {
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         Theme(this).changeTheme()
         super.onCreate(savedInstanceState)
         addPreferencesFromResource(R.xml.pref)
-
-
     }
+
     override fun onSharedPreferenceChanged(sharedPreferences: SharedPreferences?, key: String?) {
-        if (key.equals("theme")){
+        if (key.equals("theme")) {
             val mStartActivity = Intent(this, MainActivity::class.java)
             val mPendingIntentId = 123456
             val mPendingIntent = PendingIntent.getActivity(this, mPendingIntentId, mStartActivity, PendingIntent.FLAG_CANCEL_CURRENT)
