+++ app/src/main/res/drawable/pause.png
