+++ app/src/main/res/layout/activity_main.xml
@@ -6,11 +6,9 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    
     <android.support.v7.widget.RecyclerView
+        android:id="@+id/rv_songs"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/rv_songs"
         tools:listitem="@layout/card_music" />
-
 </FrameLayout>
