+++ app/src/main/res/layout/activity_music.xml
@@ -7,42 +7,6 @@
     android:background="@drawable/background"
     tools:context=".MusicActivity">
 
-    <Button
-        android:id="@+id/btn_prev"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignTop="@+id/btn_play"
-        android:layout_marginStart="19dp"
-        android:text="prev" />
-
-    <Button
-        android:id="@+id/btn_play"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="78dp"
-        android:text="play" />
-
-    <Button
-        android:id="@+id/btn_next"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignTop="@+id/btn_play"
-        android:layout_marginEnd="23dp"
-        android:text="next" />
-
-    <Button
-        android:id="@+id/btn_pause"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="16dp"
-        android:text="pause" />
-
     <ImageView
         android:id="@+id/image_song2"
         android:layout_width="250dp"
@@ -59,10 +23,10 @@
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
         android:layout_marginTop="19dp"
-        android:textStyle="bold"
-        android:textSize="25sp"
         android:text="Singer"
-        android:textColor="@color/colorSinger"/>
+        android:textColor="@color/colorSinger"
+        android:textSize="30sp"
+        android:textStyle="bold" />
 
     <TextView
         android:id="@+id/txt_song2"
@@ -72,33 +36,35 @@
         android:layout_centerHorizontal="true"
         android:layout_marginBottom="175dp"
         android:text="Smells like teen spirit"
-        android:textSize="20sp"
         android:textColor="@color/colorSong"
+        android:textSize="25sp"
         android:textStyle="italic" />
-
-    <ImageView
-        android:id="@+id/image_next"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignStart="@+id/btn_next"
-        android:layout_marginBottom="54dp"
-        app:srcCompat="@drawable/next" />
 
     <ImageView
         android:id="@+id/image_play"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_above="@+id/btn_pause"
+        android:layout_alignParentBottom="true"
         android:layout_centerHorizontal="true"
-        app:srcCompat="@drawable/play" />
+        android:layout_marginBottom="45dp"
+        app:srcCompat="@drawable/pause" />
 
     <ImageView
         android:id="@+id/image_prev"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignStart="@+id/btn_prev"
-        android:layout_alignTop="@+id/image_next"
+        android:layout_alignParentStart="true"
+        android:layout_alignTop="@+id/image_play"
+        android:layout_marginStart="19dp"
         app:srcCompat="@drawable/prev" />
+
+    <ImageView
+        android:id="@+id/image_next"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentEnd="true"
+        android:layout_alignTop="@+id/image_play"
+        android:layout_marginEnd="23dp"
+        app:srcCompat="@drawable/next" />
 
 </RelativeLayout>
