+++ app/src/main/res/layout/activity_settings.xml
@@ -9,6 +9,6 @@
     <FrameLayout
         android:id="@+id/frame"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content" />
 
 </android.support.constraint.ConstraintLayout>
