+++ app/src/main/res/values/styles.xml
@@ -11,7 +11,6 @@
     </style>
 
     <style name="AppTheme2" parent="Base.Theme.AppCompat.Light.DarkActionBar">
-        <!-- Customize your theme here. -->
         <item name="colorPrimary">@color/colorPrimaryTheme2</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDarkTheme2</item>
         <item name="colorAccent">@color/colorAccentTheme2</item>
@@ -20,7 +19,6 @@
     </style>
 
     <style name="AppTheme3" parent="Base.Theme.AppCompat.Light.DarkActionBar">
-        <!-- Customize your theme here. -->
         <item name="colorPrimary">@color/colorPrimaryTheme3</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDarkTheme3</item>
         <item name="colorAccent">@color/colorAccentTheme3</item>
@@ -29,7 +27,6 @@
     </style>
 
     <style name="AppTheme4" parent="Base.Theme.AppCompat.Light.DarkActionBar">
-        <!-- Customize your theme here. -->
         <item name="colorPrimary">@color/colorPrimaryTheme4</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDarkTheme4</item>
         <item name="colorAccent">@color/colorAccentTheme4</item>
