+++ app/src/main/res/xml/pref.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
     <ListPreference
         android:defaultValue="-1"
         android:entries="@array/pref_example_list_titles"
